Fix DisplayLocking ShadowDOM Range getBoundingClientRect crash
The traversal I am replacing in this patch is supposed to match the way
that Range traverses the DOM tree, but it doesn't. Range uses
NodeTraversal::Next, but this code was using FlatTreeTraversal::Next.
Due to this, it got in a state where the FlatTreeTraversal would never
get to the end node as specified by the Range, so it would run past the
end of the DOM tree and crash.

This new traversal is much more similar to the Range's traversal, but
also includes all shadow trees just to be safe and make sure we unlock
everything that might need to get unlocked.

Fixed: 1250391
Change-Id: I5107ba768d12beaeb440a73389ec90d3e8748feb