[wasm] Add test for cross-origin module sharing
This CL adds two more tests for cross-origin sharing of WebAssembly
modules. The first test tries to send a module to an iframe in a
different agent cluster. The second test tries to send a module to an
iframe in the same agent cluster, but cross-origin from
http://web-platform.test:8001 to
receiver: http://www1.web-platform.test:8001. In both cases the
postMessage should fail.

R=jbroman@chromium.org

Bug: chromium:1225641
Change-Id: Id1046cab521f65b68ab15bf55faee54cda70160e